+++ scripts/sidedock.py
@@ -1,6 +1,7 @@
 from compas.geometry import Box
 from compas_viewer import Viewer
 from compas_viewer.components import Button
+from compas_viewer.components import Slider
 
 viewer = Viewer()
 
@@ -12,7 +13,10 @@ def toggle_box():
     boxobj.is_visible = not boxobj.is_visible
     boxobj.update()
 
+def slider_changed(value):
+    print(f"Slider value changed to: {value}")
 
 viewer.ui.sidedock.show = True
 viewer.ui.sidedock.add(Button(text="Toggle Box", action=toggle_box))
+viewer.ui.sidedock.add(Slider(title="test", min_val=0, max_val=10, action=slider_changed))
 viewer.show()